expense_tracker_backend/requirements.txt
absl-py==1.4.0
accelerate==1.8.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.4
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
Automat==25.4.16
babel==2.17.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.3.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.1.0
chromadb==1.0.15
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.1.3
constantly==23.10.4
contourpy==1.1.0
cryptography==45.0.5
cssselect==1.3.0
curl_cffi==0.13.0
cycler==0.11.0
databricks-sdk==0.63.0
dataclasses-json==0.6.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
deta==1.2.0
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
docker==7.1.0
durationpy==0.10
executing==1.2.0
faiss-cpu==1.11.0
fastapi==0.115.14
fastjsonschema==2.21.1
ffmpy==0.6.0
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
flatbuffers==23.5.26
fonttools==4.40.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.5.1
gast==0.4.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gradio==5.34.2
gradio_client==1.10.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
groovy==0.1.2
groq==0.30.0
grpcio==1.73.1
grpcio-status==1.73.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.9.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
imbalanced-learn==0.14.0
imblearn==0.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==24.7.2
ipykernel==6.23.2
ipython==8.14.0
ipywidgets==8.1.5
isoduration==20.11.0
itemadapter==0.12.0
itemloaders==1.3.2
itsdangerous==2.2.0
jax==0.4.13
jedi==0.18.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.2.0
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==2.12.0
kiwisolver==1.4.4
kubernetes==33.1.0
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.72
langchain-google-genai==2.1.5
langchain-groq==0.3.6
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.8
langsmith==0.4.3
lazypredict==0.2.16
libclang==16.0.0
lightgbm==4.6.0
llvmlite==0.44.0
lxml==6.0.0
Mako==1.3.10
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.26.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.1.3
ml-dtypes==0.2.0
mlflow==3.2.0
mlflow-skinny==3.2.0
mlflow-tracing==3.2.0
mmh3==5.1.0
mpmath==1.3.0
multidict==6.6.0
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==1.48.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.5.6
networkx==3.5
notebook==7.3.3
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.22.1
openai==1.92.2
opencv-python==4.7.0.72
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opt-einsum==3.3.0
optuna==4.5.0
orjson==3.10.18
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.0.2
pandas-ta==0.3.14b0
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.3
patsy==1.0.1
peewee==3.18.2
pickleshare==0.7.5
Pillow==9.5.0
pipp==0.0.1
platformdirs==3.7.0
posthog==5.4.0
prometheus_client==0.21.1
prompt-toolkit==3.0.38
propcache==0.3.2
Protego==0.5.0
proto-plus==1.26.1
protobuf==6.31.1
psutil==5.9.5
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyarrow==21.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
PyDispatcher==2.0.7
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.0.9
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
pytest-runner==6.0.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2023.3
PyWavelets==1.9.0
pywin32==306
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==25.1.0
queuelib==1.8.0
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.23.1
rsa==4.9
ruff==0.12.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.16.1
Scrapy==2.13.3
seaborn==0.12.2
selenium==4.34.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==4.1.0
service-identity==24.2.0
shellingham==1.5.4
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.2
starlette==0.46.2
statsmodels==0.14.5
streamlit==1.47.0
stumpy==1.13.0
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.18.1
threadpoolctl==3.1.0
tiktoken==0.9.0
tinycss2==1.4.0
tldextract==5.3.0
tokenizers==0.21.2
toml==0.10.2
tomlkit==0.13.3
torch==2.6.0
tornado==6.3.2
tqdm==4.67.1
traitlets==5.9.0
transformers==4.53.0
trio==0.30.0
trio-websocket==0.12.2
tsfresh==0.21.0
Twisted==25.5.0
typer==0.16.0
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.34.3
w3lib==2.3.1
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.6
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.9.0
widgetsnbextension==4.0.13
wrapt==1.14.1
wsproto==1.2.0
xgboost==1.7.6
xlwt==1.3.0
yarl==1.20.1
yfinance==0.2.65
youtube-transcript-api==1.1.0
zipp==3.23.0
zope.interface==7.2
zstandard==0.23.0
